refactor: move PHPCS version compatibility check to `compatibility.ts`.
- Moved `checkPhpcsVersionCompatibility` and `getPhpcsVersion` method from `settings.ts` into it's own `compatibility.ts` file, for better organisation. Also added it's file import to `settings.ts`.

- Refactored the checks for the unsupported 4.x version in `checkPhpcsVersionCompatibility` function into a new `checkUnsupportedVersion` function, so there's one source of truth for defining unsupported versions. This is in anticipation of eventually dropping support of 3.x and supporting 4.x.
